Move docker coverage fixtures back into gitlab conftest
